use nanovg as backend

+++ src/primitives.rs
@@ -12,58 +12,58 @@ pub use draw::Path;
 
 #[derive(Clone, Debug)]
 pub struct Circle{
-    pub r: f64,
-    pub cx: f64,
-    pub cy: f64,
+    pub r: f32,
+    pub cx: f32,
+    pub cy: f32,
 }
 
 #[derive(Clone, Debug)]
 pub struct Ellipse{
-    pub rx: f64,
-    pub ry: f64,
-    pub cx: f64,
-    pub cy: f64,
+    pub rx: f32,
+    pub ry: f32,
+    pub cx: f32,
+    pub cy: f32,
 }
 
 #[derive(Clone, Debug)]
 pub struct Line{
-    pub x1: f64,
-    pub y1: f64,
-    pub x2: f64,
-    pub y2: f64,
+    pub x1: f32,
+    pub y1: f32,
+    pub x2: f32,
+    pub y2: f32,
 }
 
 #[derive(Clone, Debug)]
 pub struct Image{
-    pub x: f64,
-    pub y: f64,
-    pub height: f64,
-    pub width: f64,
+    pub x: f32,
+    pub y: f32,
+    pub height: f32,
+    pub width: f32,
     //TODO: add the data
 }
 
 #[derive(Clone, Debug)]
 pub struct Polygon{
-    pub cords: Vec<(f64, f64)>,
+    pub cords: Vec<(f32, f32)>,
 }
 
 #[derive(Clone, Debug)]
 pub struct Polyline{
-    pub cords: Vec<(f64,f64)>,
+    pub cords: Vec<(f32,f32)>,
 }
 
 #[derive(Clone, Debug)]
 pub struct Rect{
-    pub x: f64,
-    pub y: f64,
-    pub width: f64,
-    pub height: f64,
+    pub x: f32,
+    pub y: f32,
+    pub width: f32,
+    pub height: f32,
     //maybe also implement rounden corners
 }
 
 #[derive(Clone, Debug)]
 pub struct Text<'a>{
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
     pub text: &'a str,
 }
